requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
loguru==0.7.3
fastapi[standard]
passlib
python-jose
prometheus_client
bcrypt
psutil-7.1.3